refactor src directory structure. Extract pages from components folder, restricting components to reusable pieces from which pages will be composed
